handle log filter
