+++ src/io/flutter/logging/FlutterLogTree.java
@@ -19,6 +19,7 @@ import com.intellij.util.Alarm;
 import com.intellij.util.EventDispatcher;
 import com.intellij.util.ui.ColumnInfo;
 import io.flutter.console.FlutterConsoleFilter;
+import io.flutter.logging.v2.EntryFilter;
 import io.flutter.run.daemon.FlutterApp;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -350,63 +351,6 @@ public class FlutterLogTree extends TreeTable {
           append(text.substring(cursor), SimpleTextAttributes.REGULAR_ATTRIBUTES);
         }
       }
-    }
-  }
-
-  public static class EntryFilter {
-    @Nullable
-    private final String text;
-    private final boolean isRegex;
-    private final boolean isMatchCase;
-
-    public EntryFilter(@Nullable String text) {
-      this(text, false, false);
-    }
-
-    public EntryFilter(@Nullable String text, boolean isMatchCase, boolean isRegex) {
-      this.text = text;
-      this.isMatchCase = isMatchCase;
-      this.isRegex = isRegex;
-    }
-
-    @Nullable
-    public String getText() {
-      return text;
-    }
-
-    public boolean accept(@NotNull FlutterLogEntry entry) {
-      if (text == null) {
-        return true;
-      }
-      final String standardText = isMatchCase ? text : text.toLowerCase();
-      final String standardMessage = isMatchCase ? entry.getMessage() : entry.getMessage().toLowerCase();
-      final String standardCategory = isMatchCase ? entry.getCategory() : entry.getCategory().toLowerCase();
-      if (acceptByCheckingRegexOption(standardCategory, standardText)) {
-        return true;
-      }
-      return acceptByCheckingRegexOption(standardMessage, standardText);
-    }
-
-    private boolean acceptByCheckingRegexOption(@NotNull String message, @NotNull String text) {
-      if (isRegex) {
-        return message.matches("(?s).*" + text + ".*");
-      }
-      return message.contains(text);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-      if (this == o) return true;
-      if (o == null || getClass() != o.getClass()) return false;
-      final EntryFilter filter = (EntryFilter)o;
-      return isRegex == filter.isRegex &&
-             isMatchCase == filter.isMatchCase &&
-             Objects.equals(text, filter.text);
-    }
-
-    @Override
-    public int hashCode() {
-      return Objects.hash(text, isRegex, isMatchCase);
     }
   }
